+++ go.mod
@@ -23,7 +23,6 @@ require (
 	github.com/stellar/go v0.0.0-20250417130535-10df2019148f
 	github.com/stellar/go-xdr v0.0.0-20231122183749-b53fb00bcac2
 	github.com/stretchr/testify v1.10.0
-	golang.org/x/exp v0.0.0-20231006140011-7918f672742d
 	golang.org/x/term v0.31.0
 	golang.org/x/text v0.24.0
 )
@@ -67,6 +66,7 @@ require (
 	github.com/subosito/gotenv v1.6.0 // indirect
 	go.uber.org/multierr v1.11.0 // indirect
 	golang.org/x/crypto v0.37.0 // indirect
+	golang.org/x/exp v0.0.0-20231006140011-7918f672742d // indirect
 	golang.org/x/net v0.38.0 // indirect
 	golang.org/x/sys v0.32.0 // indirect
 	google.golang.org/protobuf v1.36.5 // indirect
