+++ internal/services/ingest.go
@@ -15,10 +15,8 @@ import (
 	"github.com/stellar/wallet-backend/internal/db"
 	"github.com/stellar/wallet-backend/internal/entities"
 	"github.com/stellar/wallet-backend/internal/metrics"
-	"github.com/stellar/wallet-backend/internal/tss"
-	tssrouter "github.com/stellar/wallet-backend/internal/tss/router"
-	tssstore "github.com/stellar/wallet-backend/internal/tss/store"
 	"github.com/stellar/wallet-backend/internal/utils"
+	txutils "github.com/stellar/wallet-backend/internal/transactions/utils"
 )
 
 const (
@@ -41,8 +39,6 @@ type ingestService struct {
 	ledgerCursorName string
 	appTracker       apptracker.AppTracker
 	rpcService       RPCService
-	tssRouter        tssrouter.Router
-	tssStore         tssstore.Store
 	metricsService   metrics.MetricsService
 }
 
@@ -51,8 +47,6 @@ func NewIngestService(
 	ledgerCursorName string,
 	appTracker apptracker.AppTracker,
 	rpcService RPCService,
-	tssRouter tssrouter.Router,
-	tssStore tssstore.Store,
 	metricsService metrics.MetricsService,
 ) (*ingestService, error) {
 	if models == nil {
@@ -67,12 +61,6 @@ func NewIngestService(
 	if rpcService == nil {
 		return nil, errors.New("rpcService cannot be nil")
 	}
-	if tssRouter == nil {
-		return nil, errors.New("tssRouter cannot be nil")
-	}
-	if tssStore == nil {
-		return nil, errors.New("tssStore cannot be nil")
-	}
 	if metricsService == nil {
 		return nil, errors.New("metricsService cannot be nil")
 	}
@@ -82,8 +70,6 @@ func NewIngestService(
 		ledgerCursorName: ledgerCursorName,
 		appTracker:       appTracker,
 		rpcService:       rpcService,
-		tssRouter:        tssRouter,
-		tssStore:         tssStore,
 		metricsService:   metricsService,
 	}, nil
 }
@@ -137,13 +123,6 @@ func (m *ingestService) Run(ctx context.Context, startLedger uint32, endLedger u
 				return fmt.Errorf("error ingesting payments: %w", err)
 			}
 			m.metricsService.ObserveIngestionDuration(paymentPrometheusLabel, time.Since(startTime).Seconds())
-
-			startTime = time.Now()
-			err = m.processTSSTransactions(ctx, ledgerTransactions)
-			if err != nil {
-				return fmt.Errorf("error processing tss transactions: %w", err)
-			}
-			m.metricsService.ObserveIngestionDuration(tssPrometheusLabel, time.Since(startTime).Seconds())
 
 			err = m.models.Payments.UpdateLatestLedgerSynced(ctx, m.ledgerCursorName, ingestLedger)
 			if err != nil {
@@ -201,7 +180,7 @@ func (m *ingestService) ingestPayments(ctx context.Context, ledgerTransactions [
 			if err != nil {
 				return fmt.Errorf("generic transaction cannot be unpacked into a transaction")
 			}
-			txResultXDR, err := tss.UnmarshallTransactionResultXDR(tx.ResultXDR)
+			txResultXDR, err := txutils.UnmarshallTransactionResultXDR(tx.ResultXDR)
 			if err != nil {
 				return fmt.Errorf("cannot unmarshal transacation result xdr: %s", err.Error())
 			}
@@ -252,78 +231,6 @@ func (m *ingestService) ingestPayments(ctx context.Context, ledgerTransactions [
 		return fmt.Errorf("ingesting payments: %w", err)
 	}
 
-	return nil
-}
-
-func (m *ingestService) processTSSTransactions(ctx context.Context, ledgerTransactions []entities.Transaction) error {
-	// Initialize a map to track counts by status
-	statusCounts := make(map[string]float64)
-
-	for _, tx := range ledgerTransactions {
-		if !tx.FeeBump {
-			// because all transactions submitted by TSS are fee bump transactions
-			continue
-		}
-		tssTry, err := m.tssStore.GetTry(ctx, tx.Hash)
-		if err != nil {
-			return fmt.Errorf("error when getting try: %w", err)
-		}
-		if tssTry == (tssstore.Try{}) {
-			continue
-		}
-
-		transaction, err := m.tssStore.GetTransaction(ctx, tssTry.OrigTxHash)
-		if err != nil {
-			return fmt.Errorf("error getting transaction: %w", err)
-		}
-		status := tss.RPCTXStatus{RPCStatus: tx.Status}
-		code, err := tss.TransactionResultXDRToCode(tx.ResultXDR)
-		if err != nil {
-			return fmt.Errorf("error unmarshaling resultxdr: %w", err)
-		}
-		err = m.tssStore.UpsertTry(ctx, tssTry.OrigTxHash, tssTry.Hash, tssTry.XDR, status, code, tx.ResultXDR)
-		if err != nil {
-			return fmt.Errorf("error updating try: %w", err)
-		}
-		err = m.tssStore.UpsertTransaction(ctx, transaction.WebhookURL, tssTry.OrigTxHash, transaction.XDR, status)
-		if err != nil {
-			return fmt.Errorf("error updating transaction: %w", err)
-		}
-
-		txCode, err := tss.TransactionResultXDRToCode(tx.ResultXDR)
-		if err != nil {
-			return fmt.Errorf("unable to extract tx code from result xdr string: %w", err)
-		}
-
-		tssGetIngestResponse := tss.RPCGetIngestTxResponse{
-			Status:      tx.Status,
-			Code:        txCode,
-			EnvelopeXDR: tx.EnvelopeXDR,
-			ResultXDR:   tx.ResultXDR,
-			CreatedAt:   int64(tx.CreatedAt),
-		}
-		payload := tss.Payload{
-			RPCGetIngestTxResponse: tssGetIngestResponse,
-		}
-		err = m.tssRouter.Route(payload)
-		if err != nil {
-			return fmt.Errorf("unable to route payload: %w", err)
-		}
-		// Record the transaction status transition
-		m.metricsService.RecordTSSTransactionStatusTransition(transaction.Status, string(tx.Status))
-
-		// Calculate and record the transaction inclusion time using the transaction's creation time in our system
-		inclusionTime := time.Since(transaction.CreatedAt).Seconds()
-		m.metricsService.ObserveTSSTransactionInclusionTime(string(tx.Status), inclusionTime)
-
-		// Increment the count for this status
-		statusCounts[string(tx.Status)]++
-	}
-
-	// Set the final counts for each status
-	for status, count := range statusCounts {
-		m.metricsService.SetNumTssTransactionsIngestedPerLedger(status, count)
-	}
 	return nil
 }
