+++ internal/services/ingest_test.go
@@ -22,181 +22,7 @@ import (
 	"github.com/stellar/wallet-backend/internal/db/dbtest"
 	"github.com/stellar/wallet-backend/internal/entities"
 	"github.com/stellar/wallet-backend/internal/metrics"
-	"github.com/stellar/wallet-backend/internal/tss"
-	tssrouter "github.com/stellar/wallet-backend/internal/tss/router"
-	tssstore "github.com/stellar/wallet-backend/internal/tss/store"
 )
-
-func TestGetLedgerTransactions(t *testing.T) {
-	dbt := dbtest.Open(t)
-	defer dbt.Close()
-
-	dbConnectionPool, err := db.OpenDBConnectionPool(dbt.DSN)
-	require.NoError(t, err)
-	defer dbConnectionPool.Close()
-
-	mockMetricsService := metrics.NewMockMetricsService()
-	models, err := data.NewModels(dbConnectionPool, mockMetricsService)
-	require.NoError(t, err)
-	mockAppTracker := apptracker.MockAppTracker{}
-	mockRPCService := RPCServiceMock{}
-	mockRouter := tssrouter.MockRouter{}
-	tssStore, err := tssstore.NewStore(dbConnectionPool, mockMetricsService)
-	require.NoError(t, err)
-	ingestService, err := NewIngestService(models, "ingestionLedger", &mockAppTracker, &mockRPCService, &mockRouter, tssStore, mockMetricsService)
-	require.NoError(t, err)
-	t.Run("all_ledger_transactions_in_single_gettransactions_call", func(t *testing.T) {
-		defer mockMetricsService.AssertExpectations(t)
-
-		rpcGetTransactionsResult := entities.RPCGetTransactionsResult{
-			Cursor: "51",
-			Transactions: []entities.Transaction{
-				{
-					Status: entities.SuccessStatus,
-					Hash:   "hash1",
-					Ledger: 1,
-				},
-				{
-					Status: entities.FailedStatus,
-					Hash:   "hash2",
-					Ledger: 2,
-				},
-			},
-		}
-		mockRPCService.
-			On("GetTransactions", int64(1), "", 50).
-			Return(rpcGetTransactionsResult, nil).
-			Once()
-
-		txns, err := ingestService.GetLedgerTransactions(1)
-		assert.Equal(t, 1, len(txns))
-		assert.Equal(t, txns[0].Hash, "hash1")
-		assert.NoError(t, err)
-	})
-
-	t.Run("ledger_transactions_split_between_multiple_gettransactions_calls", func(t *testing.T) {
-		defer mockMetricsService.AssertExpectations(t)
-
-		rpcGetTransactionsResult1 := entities.RPCGetTransactionsResult{
-			Cursor: "51",
-			Transactions: []entities.Transaction{
-				{
-					Status: entities.SuccessStatus,
-					Hash:   "hash1",
-					Ledger: 1,
-				},
-				{
-					Status: entities.FailedStatus,
-					Hash:   "hash2",
-					Ledger: 1,
-				},
-			},
-		}
-		rpcGetTransactionsResult2 := entities.RPCGetTransactionsResult{
-			Cursor: "51",
-			Transactions: []entities.Transaction{
-				{
-					Status: entities.SuccessStatus,
-					Hash:   "hash3",
-					Ledger: 1,
-				},
-				{
-					Status: entities.FailedStatus,
-					Hash:   "hash4",
-					Ledger: 2,
-				},
-			},
-		}
-
-		mockRPCService.
-			On("GetTransactions", int64(1), "", 50).
-			Return(rpcGetTransactionsResult1, nil).
-			Once()
-
-		mockRPCService.
-			On("GetTransactions", int64(1), "51", 50).
-			Return(rpcGetTransactionsResult2, nil).
-			Once()
-
-		txns, err := ingestService.GetLedgerTransactions(1)
-		assert.Equal(t, 3, len(txns))
-		assert.Equal(t, txns[0].Hash, "hash1")
-		assert.Equal(t, txns[1].Hash, "hash2")
-		assert.Equal(t, txns[2].Hash, "hash3")
-		assert.NoError(t, err)
-	})
-}
-
-func TestProcessTSSTransactions(t *testing.T) {
-	dbt := dbtest.Open(t)
-	defer dbt.Close()
-
-	dbConnectionPool, err := db.OpenDBConnectionPool(dbt.DSN)
-	require.NoError(t, err)
-	defer dbConnectionPool.Close()
-
-	mockMetricsService := metrics.NewMockMetricsService()
-	models, err := data.NewModels(dbConnectionPool, mockMetricsService)
-	require.NoError(t, err)
-	mockAppTracker := apptracker.MockAppTracker{}
-	mockRPCService := RPCServiceMock{}
-	mockRouter := tssrouter.MockRouter{}
-	tssStore, err := tssstore.NewStore(dbConnectionPool, mockMetricsService)
-	require.NoError(t, err)
-	ingestService, err := NewIngestService(models, "ingestionLedger", &mockAppTracker, &mockRPCService, &mockRouter, tssStore, mockMetricsService)
-	require.NoError(t, err)
-
-	t.Run("routes_to_tss_router", func(t *testing.T) {
-		mockMetricsService.On("ObserveDBQueryDuration", "INSERT", "tss_transactions", mock.AnythingOfType("float64")).Times(2)
-		mockMetricsService.On("IncDBQuery", "INSERT", "tss_transactions").Times(2)
-		mockMetricsService.On("ObserveDBQueryDuration", "INSERT", "tss_transaction_submission_tries", mock.AnythingOfType("float64")).Times(2)
-		mockMetricsService.On("IncDBQuery", "INSERT", "tss_transaction_submission_tries").Times(2)
-		mockMetricsService.On("ObserveDBQueryDuration", "SELECT", "tss_transaction_submission_tries", mock.AnythingOfType("float64")).Times(2)
-		mockMetricsService.On("IncDBQuery", "SELECT", "tss_transaction_submission_tries").Times(2)
-		mockMetricsService.On("ObserveDBQueryDuration", "SELECT", "tss_transactions", mock.AnythingOfType("float64")).Times(2)
-		mockMetricsService.On("IncDBQuery", "SELECT", "tss_transactions").Times(2)
-		mockMetricsService.On("RecordTSSTransactionStatusTransition", "NEW", "SUCCESS").Once()
-		mockMetricsService.On("ObserveTSSTransactionInclusionTime", "SUCCESS", mock.AnythingOfType("float64")).Once()
-		mockMetricsService.On("SetNumTssTransactionsIngestedPerLedger", "SUCCESS", float64(1)).Once()
-		defer mockMetricsService.AssertExpectations(t)
-
-		transactions := []entities.Transaction{
-			{
-				Status:              entities.SuccessStatus,
-				Hash:                "feebumphash",
-				ApplicationOrder:    1,
-				FeeBump:             true,
-				EnvelopeXDR:         "feebumpxdr",
-				ResultXDR:           "AAAAAAAAAMj////9AAAAAA==",
-				ResultMetaXDR:       "meta",
-				Ledger:              123456,
-				DiagnosticEventsXDR: []string{"diag"},
-				CreatedAt:           1695939098,
-			},
-		}
-
-		err = tssStore.UpsertTransaction(context.Background(), "localhost:8000/webhook", "hash", "xdr", tss.RPCTXStatus{OtherStatus: tss.NewStatus})
-		require.NoError(t, err)
-		err = tssStore.UpsertTry(context.Background(), "hash", "feebumphash", "feebumpxdr", tss.RPCTXStatus{OtherStatus: tss.NewStatus}, tss.RPCTXCode{OtherCodes: tss.NewCode}, "")
-		require.NoError(t, err)
-
-		mockRouter.
-			On("Route", mock.AnythingOfType("tss.Payload")).
-			Return(nil).
-			Once()
-
-		err := ingestService.processTSSTransactions(context.Background(), transactions)
-		assert.NoError(t, err)
-
-		updatedTX, err := tssStore.GetTransaction(context.Background(), "hash")
-		require.NoError(t, err)
-		assert.Equal(t, string(entities.SuccessStatus), updatedTX.Status)
-		updatedTry, err := tssStore.GetTry(context.Background(), "feebumphash")
-		require.NoError(t, err)
-		assert.Equal(t, "AAAAAAAAAMj////9AAAAAA==", updatedTry.ResultXDR)
-		assert.Equal(t, int32(xdr.TransactionResultCodeTxTooLate), updatedTry.Code)
-	})
-}
 
 func TestIngestPayments(t *testing.T) {
 	dbt := dbtest.Open(t)
@@ -211,10 +37,7 @@ func TestIngestPayments(t *testing.T) {
 	require.NoError(t, err)
 	mockAppTracker := apptracker.MockAppTracker{}
 	mockRPCService := RPCServiceMock{}
-	mockRouter := tssrouter.MockRouter{}
-	tssStore, err := tssstore.NewStore(dbConnectionPool, mockMetricsService)
-	require.NoError(t, err)
-	ingestService, err := NewIngestService(models, "ingestionLedger", &mockAppTracker, &mockRPCService, &mockRouter, tssStore, mockMetricsService)
+	ingestService, err := NewIngestService(models, "ingestionLedger", &mockAppTracker, &mockRPCService, mockMetricsService)
 	require.NoError(t, err)
 	srcAccount := keypair.MustRandom().Address()
 	destAccount := keypair.MustRandom().Address()
@@ -461,12 +284,8 @@ func TestIngest_LatestSyncedLedgerBehindRPC(t *testing.T) {
 	mockAppTracker := apptracker.MockAppTracker{}
 	mockRPCService := RPCServiceMock{}
 	mockRPCService.On("TrackRPCServiceHealth", ctx, mock.Anything).Once()
-	mockRouter := tssrouter.MockRouter{}
 
-	tssStore, err := tssstore.NewStore(dbConnectionPool, mockMetricsService)
-	require.NoError(t, err)
-
-	ingestService, err := NewIngestService(models, "ingestionLedger", &mockAppTracker, &mockRPCService, &mockRouter, tssStore, mockMetricsService)
+	ingestService, err := NewIngestService(models, "ingestionLedger", &mockAppTracker, &mockRPCService, mockMetricsService)
 	require.NoError(t, err)
 
 	srcAccount := keypair.MustRandom().Address()
@@ -550,12 +369,8 @@ func TestIngest_LatestSyncedLedgerAheadOfRPC(t *testing.T) {
 	mockAppTracker := apptracker.MockAppTracker{}
 	mockRPCService := RPCServiceMock{}
 	mockRPCService.On("TrackRPCServiceHealth", ctx, mock.Anything).Once()
-	mockRouter := tssrouter.MockRouter{}
 
-	tssStore, err := tssstore.NewStore(dbConnectionPool, mockMetricsService)
-	require.NoError(t, err)
-
-	ingestService, err := NewIngestService(models, "ingestionLedger", &mockAppTracker, &mockRPCService, &mockRouter, tssStore, mockMetricsService)
+	ingestService, err := NewIngestService(models, "ingestionLedger", &mockAppTracker, &mockRPCService, mockMetricsService)
 	require.NoError(t, err)
 
 	mockMetricsService.On("ObserveDBQueryDuration", "INSERT", "ingest_store", mock.AnythingOfType("float64")).Once()
