--- internal/tss/channels/types.go
@@ -1,6 +1,0 @@
-package channels
-
-type WorkerPool interface {
-	Submit(task func())
-	StopAndWait()
-}
